fix sqlite primary key constraint error when creating multiple sessions with the same profile
Multiple sessions should be able to be created for the same user profile provided
by the OIDC provider, so a random id string will be used for the profile id instead
of the user id, eliminating primary key conflicts. The user id is not required
by the app anyways since each oauth2 access token is bound to the correct resource
owner.